Rename some python script for standardization
